+++ src/nodelets/BosonCamera.cpp
@@ -26,6 +26,11 @@ PLUGINLIB_EXPORT_CLASS(flir_boson_usb::BosonCamera, nodelet::Nodelet)
 using namespace cv;
 using namespace flir_boson_usb;
 
+BosonCamera::~BosonCamera()
+{
+  closeCamera();
+}
+
 void BosonCamera::onInit()
 {
   ros::NodeHandle nh = getNodeHandle();
@@ -92,14 +97,18 @@ void BosonCamera::onInit()
     ROS_ERROR("flir_boson_usb - Invalid sensor_type value provided. Exiting.");
   }
 
+  exit = !openCamera();
+
   if (exit)
   {
     ros::shutdown();
     return;
   }
 
+  capture_timer = nh.createTimer(ros::Duration(0.03333), boost::bind(&BosonCamera::captureAndPublish, this, _1));
+
   image_transport::ImageTransport it(nh);
-  pub = it.advertise("image", 10);
+  image_pub = it.advertise("image", 10);
 }
 
 // AGC Sample ONE: Linear from min to max.
@@ -145,5 +154,236 @@ void BosonCamera::agcBasicLinear(const Mat& input_16,
 
       output_8.at<uchar>(i, j) = static_cast<uint8_t>(value4 & 0xFF);
     }
+  }
+}
+
+bool BosonCamera::openCamera()
+{
+  // Open the Video device
+  if ((fd = open(dev_path.c_str(), O_RDWR)) < 0)
+  {
+    ROS_ERROR("flir_boson_usb - ERROR : OPEN. Invalid Video Device.");
+    return false;
+  }
+
+  // Check VideoCapture mode is available
+  if (ioctl(fd, VIDIOC_QUERYCAP, &cap) < 0)
+  {
+    ROS_ERROR("flir_boson_usb - ERROR : VIDIOC_QUERYCAP. Video Capture is not available.");
+    return false;
+  }
+
+  if (!(cap.capabilities & V4L2_CAP_VIDEO_CAPTURE))
+  {
+    ROS_ERROR("flir_boson_usb - The device does not handle single-planar video capture.");
+    return false;
+  }
+
+  struct v4l2_format format;
+
+  // Two different FORMAT modes, 8 bits vs RAW16
+  if (video_mode == RAW16)
+  {
+    // I am requiring thermal 16 bits mode
+    format.fmt.pix.pixelformat = V4L2_PIX_FMT_Y16;
+
+    // Select the frame SIZE (will depend on the type of sensor)
+    switch (sensor_type)
+    {
+      case Boson320:  // Boson320
+        width = 320;
+        height = 256;
+        break;
+      case Boson640:  // Boson640
+        width = 640;
+        height = 512;
+        break;
+      default:  // Boson320
+        width = 320;
+        height = 256;
+        break;
+    }
+  }
+  else  // 8- bits is always 640x512 (even for a Boson 320)
+  {
+    format.fmt.pix.pixelformat = V4L2_PIX_FMT_YVU420;  // thermal, works   LUMA, full Cr, full Cb
+    width = 640;
+    height = 512;
+  }
+
+  // Common varibles
+  format.type = V4L2_BUF_TYPE_VIDEO_CAPTURE;
+  format.fmt.pix.width = width;
+  format.fmt.pix.height = height;
+
+  // request desired FORMAT
+  if (ioctl(fd, VIDIOC_S_FMT, &format) < 0)
+  {
+    ROS_ERROR("flir_boson_usb - VIDIOC_S_FMT error.");
+    return false;
+  }
+
+  // we need to inform the device about buffers to use.
+  // and we need to allocate them.
+  // we'll use a single buffer, and map our memory using mmap.
+  // All this information is sent using the VIDIOC_REQBUFS call and a
+  // v4l2_requestbuffers structure:
+  struct v4l2_requestbuffers bufrequest;
+  bufrequest.type = V4L2_BUF_TYPE_VIDEO_CAPTURE;
+  bufrequest.memory = V4L2_MEMORY_MMAP;
+  bufrequest.count = 1;   // we are asking for one buffer
+
+  if (ioctl(fd, VIDIOC_REQBUFS, &bufrequest) < 0)
+  {
+    ROS_ERROR("flir_boson_usb - VIDIOC_REQBUFS error.");
+    return false;
+  }
+
+  // Now that the device knows how to provide its data,
+  // we need to ask it about the amount of memory it needs,
+  // and allocate it. This information is retrieved using the VIDIOC_QUERYBUF call,
+  // and its v4l2_buffer structure.
+
+  memset(&bufferinfo, 0, sizeof(bufferinfo));
+
+  bufferinfo.type = V4L2_BUF_TYPE_VIDEO_CAPTURE;
+  bufferinfo.memory = V4L2_MEMORY_MMAP;
+  bufferinfo.index = 0;
+
+  if (ioctl(fd, VIDIOC_QUERYBUF, &bufferinfo) < 0)
+  {
+    ROS_ERROR("flir_boson_usb - VIDIOC_QUERYBUF error.");
+    return false;
+  }
+
+  // map fd+offset into a process location (kernel will decide due to our NULL). length and
+  // properties are also passed
+  buffer_start = mmap(NULL, bufferinfo.length, PROT_READ | PROT_WRITE, MAP_SHARED, fd, bufferinfo.m.offset);
+
+  if (buffer_start == MAP_FAILED)
+  {
+    ROS_ERROR("flir_boson_usb - mmap error.");
+    return false;
+  }
+
+  // Fill this buffer with ceros. Initialization. Optional but nice to do
+  memset(buffer_start, 0, bufferinfo.length);
+
+  // Activate streaming
+  int type = bufferinfo.type;
+  if (ioctl(fd, VIDIOC_STREAMON, &type) < 0)
+  {
+    ROS_ERROR("flir_boson_usb - VIDIOC_STREAMON error.");
+    return false;
+  }
+
+  // Declarations for RAW16 representation
+  // Will be used in case we are reading RAW16 format
+  // Boson320 , Boson 640
+  // OpenCV input buffer  : Asking for all info: two bytes per pixel (RAW16)  RAW16 mode`
+  thermal16 = Mat(height, width, CV_16U, buffer_start);
+  // OpenCV output buffer : Data used to display the video
+  thermal16_linear = Mat(height, width, CV_8U, 1);
+
+  Mat thermal16_linear_zoom;   // (height,width, CV_8U, 1);    // Final representation
+  Mat thermal_rgb_zoom;   // (height,width, CV_8U, 1);    // Final representation
+
+  // Declarations for 8bits YCbCr mode
+  // Will be used in case we are reading YUV format
+  // Boson320, 640 :  4:2:0
+  int luma_height = height+height/2;
+  int luma_width = width;
+  int color_space = CV_8UC1;
+
+  // Declarations for Zoom representation
+  // Will be used or not depending on program arguments
+  thermal_luma = Mat(luma_height, luma_width,  color_space, buffer_start);  // OpenCV input buffer
+  // OpenCV output buffer , BGR -> Three color spaces :
+  // (640 - 640 - 640 : p11 p21 p31 .... / p12 p22 p32 ..../ p13 p23 p33 ...)
+  thermal_rgb = Mat(height, width, CV_8UC3, 1);
+
+  return true;
+}
+
+bool BosonCamera::closeCamera()
+{
+  // Finish loop. Exiting.
+  // Deactivate streaming
+  int type = bufferinfo.type;
+  if (ioctl(fd, VIDIOC_STREAMOFF, &type) < 0 )
+  {
+    ROS_ERROR("flir_boson_usb - VIDIOC_STREAMOFF error.");
+    return false;
+  };
+
+  close(fd);
+
+  return true;
+}
+
+void BosonCamera::captureAndPublish(const ros::TimerEvent& evt)
+{
+  Size size(640, 512);
+
+  if (video_mode == RAW16)
+  {
+    // -----------------------------
+    // RAW16 DATA
+    agcBasicLinear(thermal16, thermal16_linear, height, width);
+
+    // Display thermal after 16-bits AGC... will display an image
+    if (zoom_enable == 0)
+    {
+      // Threshold using Otsu's method, then use the result as a mask on the original image
+      Mat mask_mat, masked_img;
+      threshold(thermal16_linear, mask_mat, 0, 255, CV_THRESH_BINARY|CV_THRESH_OTSU);
+      thermal16_linear.copyTo(masked_img, mask_mat);
+
+      // Normalize the pixel values to the range [0, 1] then raise to power (gamma). Then convert back for display.
+      Mat d_out_img, norm_image, d_norm_image, gamma_corrected_image, d_gamma_corrected_image;
+      double gamma = 0.8;
+      masked_img.convertTo(d_out_img, CV_64FC1);
+      normalize(d_out_img, d_norm_image, 0, 1, NORM_MINMAX, CV_64FC1);
+      pow(d_out_img, gamma, d_gamma_corrected_image);
+      d_gamma_corrected_image.convertTo(gamma_corrected_image, CV_8UC1);
+      normalize(gamma_corrected_image, gamma_corrected_image, 0, 255, NORM_MINMAX, CV_8UC1);
+
+      // Apply top hat filter
+      int erosion_size = 5;
+      Mat top_hat_img, kernel = getStructuringElement(MORPH_ELLIPSE,
+          Size(2 * erosion_size + 1, 2 * erosion_size + 1));
+      morphologyEx(gamma_corrected_image, top_hat_img, MORPH_TOPHAT, kernel);
+
+      cv_img.image = thermal16_linear;
+      cv_img.header.stamp = ros::Time::now();
+      cv_img.header.frame_id = "boson_camera";
+      cv_img.encoding = "mono8";
+      pub_image = cv_img.toImageMsg();
+      image_pub.publish(pub_image);
+    }
+    else
+    {
+      resize(thermal16_linear, thermal16_linear_zoom, size);
+
+      cv_img.image = thermal16_linear_zoom;
+      cv_img.header.stamp = ros::Time::now();
+      cv_img.header.frame_id = "boson_camera";
+      cv_img.encoding = "mono8";
+      pub_image = cv_img.toImageMsg();
+      image_pub.publish(pub_image);
+    }
+  }
+  else  // Video is in 8 bits YUV
+  {
+    // ---------------------------------
+    // DATA in YUV
+    cvtColor(thermal_luma, thermal_rgb, COLOR_YUV2GRAY_I420, 0);
+
+    cv_img.image = thermal_rgb;
+    cv_img.encoding = "mono8";
+    cv_img.header.stamp = ros::Time::now();
+    cv_img.header.frame_id = "boson_camera";
+    pub_image = cv_img.toImageMsg();
+    image_pub.publish(pub_image);
   }
 }
